all done in catering, just need to ask A about a few item prices and desserts. Facebook link works as well

+++ app/src/main/java/watchtower/ayalacashier/CateringObjectInfo.java
@@ -35,7 +35,7 @@ public class CateringObjectInfo implements Serializable{
         amount = new Integer(a);
     }
 
-    public double getAmount()
+    public int getAmount()
     {
         return amount;
     }
